+++ client/my-sites/site-settings/settings-writing/controller.js
@@ -11,7 +11,7 @@ import React from 'react';
  */
 import WritingMain from 'my-sites/site-settings/settings-writing/main';
 import Taxonomies from 'my-sites/site-settings/taxonomies';
-import PodcastDetails from 'my-sites/site-settings/podcast-details';
+import PodcastDetails from 'my-sites/site-settings/podcasting-details';
 
 export default {
 	writing( context, next ) {
